pyproject.toml
[project]
name = "agents-sdk-test"
version = "0.1.0"
description = ""
authors = [
    {name = "kabadc",email = "34518306+kabadc@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents (>=0.2.4,<0.3.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "redis (>=6.2.0,<7.0.0)"
]

[tool.poetry]
packages = [{include = "agents_sdk_test", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
